# dockstore-webservice/src/main/resources/openapi3/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: theglobalalliance@genomicsandhealth.org
    name: Dockstore@ga4gh
    url: https://discuss.dockstore.org/t/opening-helpdesk-tickets/1506
  description: "This describes the dockstore API, a webservice that manages pairs\
    \ of Docker images and associated metadata such as CWL documents and Dockerfiles\
    \ used to build those images. Explore swagger.json for a Swagger 2.0 description\
    \ of our API and explore openapi.yaml for OpenAPI 3.0 descriptions."
  license:
    name: Apache License Version 2.0
    url: https://github.com/dockstore/dockstore/blob/develop/LICENSE
  termsOfService: TBD
  title: Dockstore API
  version: 1.11.0-alpha.2-SNAPSHOT
servers:
- description: Current server when hosted on AWS
  url: /api
  variables: {}
- description: When working locally
  url: /
  variables: {}
- description: Production server
  url: https://dockstore.org/api
  variables: {}
- description: Staging server
  url: https://staging.dockstore.org/api
  variables: {}
- description: Nightly build server
  url: https://dev.dockstore.net/api
  variables: {}
...
paths:
...
  /containers/hostedEntry:
    post:
      description: Create a hosted tool.
      operationId: createHostedTool
      parameters:
      - in: query
        name: registry
        schema:
          type: string
      - in: query
        name: name
        schema:
          type: string
      - in: query
        name: descriptorType
        schema:
          type: string
      - in: query
        name: namespace
        schema:
          type: string
      - in: query
        name: entryName
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
          description: Successfully created hosted entry
      security:
      - bearer: []
      tags:
      - hosted
  /containers/hostedEntry/{entryId}:
    delete:
      description: Delete a revision of a hosted tool.
      operationId: deleteHostedToolVersion
      parameters:
      - in: path
        name: entryId
        required: true
        schema:
          type: integer
          format: int64
      - in: query
        name: version
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
          description: Successfully deleted hosted entry version
      security:
      - bearer: []
      tags:
      - hosted
    patch:
      description: Non-idempotent operation for creating new revisions of hosted tools.
      operationId: editHostedTool
      parameters:
      - description: Entry to modify
        in: path
        name: entryId
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/SourceFile'
              uniqueItems: true
        description: "Set of updated sourcefiles, add files by adding new files with\
          \ unknown paths, delete files by including them with emptied content"
        required: true
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DockstoreTool'
          description: default response
      security:
      - bearer: []
      tags:
      - hosted
...
  /organizations/{organizationId}/collections/{collectionId}/entry:
    delete:
      description: Delete an entry to a collection.
      operationId: deleteEntryFromCollection
      parameters:
      - description: Organization ID.
        in: path
        name: organizationId
        required: true
        schema:
          type: integer
          format: int64
      - description: Collection ID.
        in: path
        name: collectionId
        required: true
        schema:
          type: integer
          format: int64
      - description: Entry ID.
        in: query
        name: entryId
        required: true
        schema:
          type: integer
          format: int64
      - description: Version ID.
        in: query
        name: versionId
        schema:
          type: string
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
          description: default response
      security:
      - bearer: []
      summary: Delete an entry to a collection.
      tags:
      - organizations
    post:
      description: Add an entry to a collection.
      operationId: addEntryToCollection
      parameters:
      - description: Organization ID.
        in: path
        name: organizationId
        required: true
        schema:
          type: integer
          format: int64
      - description: Collection ID.
        in: path
        name: collectionId
        required: true
        schema:
          type: integer
          format: int64
      - description: Entry ID.
        in: query
        name: entryId
        required: true
        schema:
          type: integer
          format: int64
      - description: Version ID.
        in: query
        name: versionId
        schema:
          type: integer
          format: int64
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
          description: default response
      security:
      - bearer: []
      summary: Add an entry to a collection.
      tags:
      - organizations
...
  /workflows/hostedEntry:
    post:
      description: Create a hosted workflow.
      operationId: createHostedWorkflow
      parameters:
      - in: query
        name: registry
        schema:
          type: string
      - in: query
        name: name
        schema:
          type: string
      - in: query
        name: descriptorType
        schema:
          type: string
      - in: query
        name: namespace
        schema:
          type: string
      - in: query
        name: entryName
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
          description: Successfully created hosted entry
      security:
      - bearer: []
      tags:
      - hosted
  /workflows/hostedEntry/{entryId}:
    delete:
      description: Delete a revision of a hosted workflow.
      operationId: deleteHostedWorkflowVersion
      parameters:
      - in: path
        name: entryId
        required: true
        schema:
          type: integer
          format: int64
      - in: query
        name: version
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
          description: Successfully deleted hosted entry version
      security:
      - bearer: []
      tags:
      - hosted
    patch:
      description: Non-idempotent operation for creating new revisions of hosted workflows
      operationId: editHostedWorkflow
      parameters:
      - description: Entry to modify
        in: path
        name: entryId
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/SourceFile'
              uniqueItems: true
        description: "Set of updated sourcefiles, add files by adding new files with\
          \ unknown paths, delete files by including them with emptied content"
        required: true
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
          description: default response
      security:
      - bearer: []
      tags:
      - hosted
    post:
      deprecated: true
      operationId: addZip
      parameters:
      - description: hosted entry ID
        in: path
        name: entryId
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
          description: successful operation
      security:
      - bearer: []
      summary: Creates a new revision of a hosted workflow from a zip
      tags:
      - hosted
...
  /workflows/path/entry/{repository}:
    get:
      description: Get an entry by path.
      operationId: getEntryByPath
      parameters:
      - in: path
        name: repository
        required: true
        schema:
          type: string
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
          description: default response
      security:
      - bearer: []
      tags:
      - workflows
...
  /workflows/{workflowId}/workflowVersions:
    get:
      description: Return first 200 versions in an entry
      operationId: getWorkflowVersions
      parameters:
      - description: id of the worflow
        in: path
        name: workflowId
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkflowVersion'
          description: Get list workflow versions in a workflow
        "400":
          description: Bad Request
      security:
      - bearer: []
      tags:
      - workflows
    put:
      description: Update the workflow versions linked to a workflow.
      operationId: updateWorkflowVersion
      parameters:
      - in: path
        name: workflowId
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          '*/*':
            schema:
              type: array
              items:
                $ref: '#/components/schemas/WorkflowVersion'
      responses:
        default:
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkflowVersion'
                uniqueItems: true
          description: default response
      security:
      - bearer: []
      tags:
      - workflows
...
components:
...
  securitySchemes:
    bearer:
      scheme: bearer
      type: http